requirements.txt
aiocontextvars==0.2.2
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
click==7.1.2
colorama==0.4.4
console-menu==0.6.0
contextvars==2.4
coverage==4.5.4
cursor==1.3.4
Deprecated==1.2.6
docopt==0.6.2
Flask==1.1.2
halo==0.0.31
idna==2.8
immutables==0.14
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.11.2
log-symbols==0.0.14
loguru==0.5.3
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
py==1.8.0
PyGithub==1.44.1
PyJWT==1.7.1
pyparsing==2.4.4
-e git+https://github.com/GeoffWilliams/pyreleaser_io.git@49804057119d84fd9820c93147d4c628e5eb63ce#egg=pyreleaser_io
pytest==5.2.2
pytest-cov==2.8.1
PyYAML==5.1.2
requests==2.22.0
six==1.15.0
spinners==0.0.24
termcolor==1.1.0
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==1.0.1
wrapt==1.11.2
zipp==0.6.0
